Fix instanciation errors of `settingsobj.Settings`.
Depending on the actual code path taken, `settingsobj.Settings.single`
may end up being `None` even when the code assumes it's not.  This
patch turns the `settingsobj.Settings` into a Borg, so that all
instances share the same state, and we can just get an instance with
`Settings()` and operate on it -- it will always be a valid `Settings`
object.

See http://code.activestate.com/recipes/66531-singleton-we-dont-need-no-stinkin-singleton-the-bo/
for a discussion and sample implementations of the Borg and Singleton
patterns.